backend/pyproject.toml
[tool.poetry]
name = "survey-platform-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kubernetes = "^30.1.0"
fastapi = "^0.112.1"
docker = "^7.1.0"
uvicorn = "^0.30.6"
gitpython = "^3.1.43"
sqlalchemy = "^2.0.35" 
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
jinja2 = "^3.1.4"
passlib = "^1.7.4"
minio = "^7.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
